fix(model): respect discriminators with `Model.validate()`
Fix #12621
