+++ lib/model.js
@@ -36,6 +36,7 @@ const assignVals = require('./helpers/populate/assignVals');
 const castBulkWrite = require('./helpers/model/castBulkWrite');
 const createPopulateQueryFilter = require('./helpers/populate/createPopulateQueryFilter');
 const getDefaultBulkwriteResult = require('./helpers/getDefaultBulkwriteResult');
+const getSchemaDiscriminatorByValue = require('./helpers/discriminator/getSchemaDiscriminatorByValue');
 const discriminator = require('./helpers/model/discriminator');
 const firstKey = require('./helpers/firstKey');
 const each = require('./helpers/each');
@@ -4474,7 +4475,11 @@ Model.validate = function validate(obj, pathsToValidate, context, callback) {
   }
 
   return this.db.base._promiseOrCallback(callback, cb => {
-    const schema = this.schema;
+    let schema = this.schema;
+    const discriminatorKey = schema.options.discriminatorKey;
+    if (schema.discriminators != null && obj != null && obj[discriminatorKey] != null) {
+      schema = getSchemaDiscriminatorByValue(schema, obj[discriminatorKey]) || schema;
+    }
     let paths = Object.keys(schema.paths);
 
     if (pathsToValidate != null) {
